[FIX] web: async operations should not live in init
Support for the "allow exports" group was implemented by checking the
group in init(), which is sync. Move that check over to willStart
instead.

Also update incorrect test: because we now do a "real" rpc when
initialising list views, a test which didn't properly wait on the
corresponding widget would carry on with an uninitialized widget.