[FIX] mail: Access rights error when opening the guidelines of a channel mail
Before this commit:

An Access rights error raised when trying to open the 'Guidelines of channel XX'
mail. This is happening because here for this mail, author_id is an environment
user partner (i.e Demo User) and recipient_ids is also 'Demo User' who doesn't
have the rights to read mail.message as there is no recipient in it. And also
users are going to send mail to themselves which should not be the case.

After this commit:

An access rights error will not be raised as now the channel creator will send
the mail regarding 'Guidelines of channel' to respective recipients and for that
set author_id as a channel creator partner instead of environment users
partner and add partner_ids while sending the mail.

LINKS

PR https://github.com/odoo/odoo/pull/54934
Task-2282599