[FIX] web: prevent dropdown closing on mouseup
before this commit,
selecting the whole content of the custom filter input and releasing the cursor
outside of the dropdown, closes the dropdown.
it happens because of the click event bind for the window click

this commit, replace the event from click to mousedown.
so dropdown should only close when there is a mousedown outside the dropdown.

task - 2276544