[ADD] core: unify o2m and m2m fields internally
Apparently I long assumed divergences in behaviour of these which
don't exist, or haven't existed in some time. As a result, the SSF
would treat o2m and m2m values extremely differently rather than just
treating o2m and m2m "widgets" as different behaviors.

This leads to issue with the recently added support for
widget=many2many:

* apparently the web client always fetches sub-views for m2m fields,
  meaning m2m subfields to get flagged as valid onchange targets, the
  SSF did not do that and m2m subfields being updated by an onchange
  would be discarded not even by the SSF but by onchange() itself
* an o2m with widget=many2many can receive "o2m commands" (e.g. 0, 1),
  meaning internally m2ms must handle these correctly, the "UI"
  behavior of m2ms (aka only adding existing values) is pretty much
  just a UI behaviour

Note: the use-case for `o2m[widget=many2many]` is basically a
"constrained many2many" where the relation is not symmetrical, but at
the same time is not an *ownership* relation (the "many" side of the
field can exist independently from the relation).

* add a test to try and surface the reported misbehavior
* onchange doesn't even return changes
* look at what the web client does
* turns out the webclient always fetches subviews for m2ms SSF only
  does that for o2ms, so only one2many[widget=many2many] *with embedded
  views* would have _onchange_spec flag their subfields, which is the
  case of the one in the report but not of the test one
* fix to always fetch and graft subviews for m2ms as well, test now
  properly fails with "Unsupported M2M command 1" as expected

Task 2313405