[FIX] web_editor: fix issue of disappearing content on click save button
issue occurs due view mode is saved as base view, not as extension view.
because if the view has an inherit_id mode of view can't change or set automatically.

before commit:
when user open website on edit mode and drag and drop snippet from
snippet panel to sidebar and click save button.changes of user is not visible
in sidebar

for ex. if user open open website in edit mode and drag and drop
Title snippet from snippet panel to Follow Us in sidebar and click
on save button website reload automatically after reload website Title
snippet is not visible in Follow Us in sidebar.

task-2311520
closes odoo/odoo#55908

Co-authored-by: jpr-odoo <jpr@openerp.com>