[IMP] calendar: improve the quality of the calendar app
- Remove organisers filter from the sidebar

- Responsible becomes Organizer

- Modify the design of the calendar event depending of
the actual user and if he is the owner of the event and also
of the invitation status. If the status needs an action from
the user or is uncertain, only the outline of the box is displayed.
If the user has accepted the invitation, the display is the same as
before. The same thing happens if the invitation is declined but the
text is also crossed and the box has a lower opacity.

Task ID 2196775

Co-authored-by: Simon Genin (ges) <ges@odoo.com>
Co-authored-by: jeh-odoo <jeh@odoo.com>