[FIX] calendar: assure that we keep the same behavior whith quickcreate
When quickcreating an event on the calendar app, the data received does
not have a partner_id. This lead to inconstintancy with the behavior of
the "normal" create. The last one, send invitiations to the attendees which
can't be done if there is no partner_ids at the creation. This also interfers
with the attendee status.

We now assure at the creation that if there is no attendee, we set the actual
user as one.

Task ID 2196775

Co-authored-by: Simon Genin (ges) <ges@odoo.com>
Co-authored-by: jeh-odoo <jeh@odoo.com>