[IMP] web: Allow only one dropdown open at same time
Before this commit, if you click outside an open the dropdown menu, the
current open dropdown menu is still open. This can be confusing for the
end user because it's not the same behavior used by the bootstrap
dropdown (like the mobile action menu). To avoid this confusing behavior
now we auto-close the dropdown menu when the user click outside the
dropdown menu like the bootstrap dropdown

Task ID: 1879566