[FIX] base: unwanted line break in report header address field
There is an unwanted line break in the company address in the report header.
The street value with some long string will breaks into two line instead of
display in the single line.

With this commit, we have used the div tag instead span tag to display the whole
string in the block format. So it will display the string in the proper format.

Task : 2321081