[IMP] web_editor: add initial weight to image quality display
Previously, when editing the quality/width of an image in the editor,
you would only see its new size, but not the previous size, meaning you
do not know how much you've gained if any. This commit changes the image
size display element so that it's more inline with other elements in the
left panel and displays what the initial size of the image was before
modifications.