[FIX] web: flicker issue on tab in the form view
before this commit,
activating the tab other than first one and then on saving time or editing time
a small flicker occurs. first, it display the first tab and then the actual tab
where the user left it.

this commit fixes the flickering issue by passing the correct condition.

task - 2321220