[IMP] web_editor: add loader for slow snippet options
This commit attempts to add a loader over the editor if an option takes
some time to be applied (>1s) to indicate that something is being done.

In this commit, the following changes have been made:

- Apply a modifier on the _mutex attribute to add "execWithLoadingEffect"
method (will be used to add actions to the mutex queue and set a loading
screen over the editor 1s later).

- As soon as the mutex is unlocked (all computations are completed),
the loading effect will be removed.

task-2213114